Extract some intf so that bits are easier to mock for testing
