+++ nexus-maven-plugins/nexus-maven-plugin/src/main/java/org/sonatype/nexus/plugin/settings/usertoken/MasterPasswordEncryption.java
@@ -12,62 +12,12 @@
  */
 package org.sonatype.nexus.plugin.settings.usertoken;
 
-import com.google.common.annotations.VisibleForTesting;
-import org.codehaus.plexus.component.annotations.Component;
-import org.codehaus.plexus.component.annotations.Requirement;
-import org.sonatype.plexus.components.cipher.DefaultPlexusCipher;
-import org.sonatype.plexus.components.sec.dispatcher.DefaultSecDispatcher;
-import org.sonatype.plexus.components.sec.dispatcher.SecDispatcher;
-import org.sonatype.plexus.components.sec.dispatcher.SecUtil;
-import org.sonatype.plexus.components.sec.dispatcher.model.SettingsSecurity;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
 /**
  * Helper to deal with master-password encryption.
  *
  * @since 2.1
  */
-@Component(role=MasterPasswordEncryption.class)
-public class MasterPasswordEncryption
+public interface MasterPasswordEncryption
 {
-    @Requirement
-    private SecDispatcher dispatcher;
-
-    // Constructor for Plexus
-    public MasterPasswordEncryption() {
-        super();
-    }
-
-    @VisibleForTesting
-    public MasterPasswordEncryption(final SecDispatcher dispatcher) {
-        this.dispatcher = checkNotNull(dispatcher);
-    }
-
-    // NOTE: This is copied from MavenCli as there is no other component which exposes this functionality :-(
-
-    public String encrypt(final String value) throws Exception {
-        String configurationFile = ((DefaultSecDispatcher)dispatcher).getConfigurationFile(); // HACK: Need the impl not the intf
-
-        if (configurationFile.startsWith("~")) {
-            configurationFile = System.getProperty("user.home") + configurationFile.substring(1);
-        }
-
-        String file = System.getProperty(DefaultSecDispatcher.SYSTEM_PROPERTY_SEC_LOCATION, configurationFile);
-
-        String master = null;
-
-        SettingsSecurity sec = SecUtil.read(file, true);
-        if (sec != null) {
-            master = sec.getMaster();
-        }
-
-        if (master == null) {
-            throw new IllegalStateException("Master password is not set in the setting security file: " + file);
-        }
-
-        DefaultPlexusCipher cipher = new DefaultPlexusCipher();
-        String masterPasswd = cipher.decryptDecorated(master, DefaultSecDispatcher.SYSTEM_PROPERTY_SEC_LOCATION);
-        return cipher.encryptAndDecorate(value, masterPasswd);
-    }
+    String encrypt(final String value) throws Exception;
 }
